java to object-c
google opensource project.
J2ObjC is an open-source command-line tool from Google that translates
Java code to Objective-C for the iOS (iPhone/iPad) platform. This tool
enables Java code to be part of an iOS application's build, as no
editing of the generated files is necessary. The goal is to write an
app's non-UI code (such as data access, or application logic) in Java,
which is then shared by web apps (using GWT), Android apps, and iOS
apps.